eventarc-agentic-workflows/services/agents/adk_mcp_agent/requirements.txt
google-adk>=1.6.1
google-cloud-eventarc-publishing
mcp
fastapi
uvicorn
google-cloud-modelarmor>=0.1.0
